requirements.txt
selenium==4.18.1
google-cloud-vision==3.5.0
Pillow==10.2.0
pyautogui==0.9.54
psutil==5.9.8
webdriver-manager>=4.0.1
